+++ src/open_refinery/__init__.py
@@ -205,7 +205,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 
 __all__ = [
     "Factory",